+++ src/client/responsive.ts
@@ -22,9 +22,6 @@ $(function() {
    const breakAt = 1024;
    // default features
    const feature: PageFeature = {
-      sideMenu: true,
-      postMenu: true,
-      twitter: true,
       facebook: false,
       timestamp: 0
    };
@@ -70,33 +67,37 @@ $(function() {
       if (mobileLoaded) {
          return;
       }
+      const imageStyle = { width: '100%', height: 'auto' };
 
       // could be optimized into a lazy-load
-      $('#mobile-menu').load('/mobile-menu', () => {
-         $.getScript('/js/mobile-menu.js');
+      $('#mobile-menu').load('/mobile-menu?t=' + feature.timestamp, () => {
+         $.getScript('/js/mobile-menu.js?t=' + feature.timestamp);
+      });
+
+      $('figure').each(function(this: HTMLElement) {
+         $(this)
+            .css(imageStyle)
+            .find('img')
+            .css(imageStyle);
       });
 
       mobileLoaded = true;
    }
 
    /**
-    * Lazy-load desktop resources
+    * Lazy-load desktop resources. Append timestamp to break caches.
     */
    function loadDesktop() {
       if (desktopLoaded) {
          return;
       }
 
-      // could optimized into a lazy-load
-      if (feature.sideMenu) {
-         $('#category-menu').load('/category-menu');
-      }
-
-      if (feature.postMenu) {
-         // append timestap to defeat caching between site deployments
-         $.getScript('/js/post-menu-data.js?t=' + feature.timestamp);
-         $.getScript('/js/post-menu.js');
-      }
+      // could be optimized into a lazy-load
+      $('#category-menu')
+         .load('/category-menu?t=' + feature.timestamp)
+         .on('change', 'select', e => {
+            window.location.assign($(e.target).val() as string);
+         });
 
       if (feature.facebook) {
          loadSource(
@@ -104,10 +105,6 @@ $(function() {
             '//connect.facebook.net/en_US/all.js#xfbml=1&appId=110860435668134',
             true
          );
-      }
-
-      if (feature.twitter) {
-         loadSource('twitter-wjs', '//platform.twitter.com/widgets.js');
       }
 
       desktopLoaded = true;