+++ src/controllers/post.ts
@@ -20,8 +20,9 @@ function send(
             render(viewName, {
                post: p,
                title: p.title,
+               scrollNav: true,
+               useFacebook: true,
                jsonLD: p.jsonLD(),
-               layout: Layout.None,
                description: p.longDescription,
                slug: key
             });