+++ src/routes.ts
@@ -129,14 +129,12 @@ function standard(app: Express.Application) {
    app.get('/geo.json', map.json.blog);
    app.get('/rss', postFeed);
    app.get('/about', staticPage.about);
-   app.get('/js/post-menu-data.js', menu.data);
    app.get('/sitemap.xml', staticPage.siteMap);
    app.get(`/exif/${photoID}`, photo.exif);
    app.get('/issues?', staticPage.issues);
    app.get('/issues?/:slug' + s, staticPage.issues);
    app.get('/category-menu', category.menu);
    app.get('/mobile-menu', menu.mobile);
-   app.get('/search', staticPage.search);
 
    app.use(`/${rootCategory}`, categoryRoutes());