+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## 5.0.0
+
+-  Change layout
+-  Simplify category selector (remove JavaScript)
+-  Remove Google Analytics (unecessary privacy invasion)
+-  Remove Facebook "Like" button
+-  Remove Twitter link and API
+
 ## 4.3.1
 
-- Update Google API and other dependencies
+-  Update Google API and other dependencies
 
 ## 4.3.0